fix: type wb context middleware
