+++ apps/backend/src/types/wb-context.d.ts
@@ -1,0 +1,22 @@
+declare global {
+  interface WbContext {
+    correlationId: string;
+    autonomyLevel?: number;
+    roleId?: string;
+    role?: string;
+    intent?: string;
+    when?: string;
+    autonomy?: number;
+    selectedId?: string;
+    selectedType?: string;
+  }
+
+  namespace Express {
+    interface Request {
+      wb?: WbContext;
+      correlationId?: string;
+    }
+  }
+}
+
+export {};
